+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Texas holdem".
 - Tests for "Solid integer".
 - Tests for "Bit count to limit".
+- Tests for "Porcupine fever".
 
 ## [1.15.0] - 2022-11-30
 ### Added